// src/plugins/email-template/ui/generated-admin-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type Node = {
  id: Scalars['ID']['output'];
};
// ...
export enum OrderType {
  Aggregate = 'Aggregate',
  Regular = 'Regular',
  Seller = 'Seller'
}
// ...
export type Seller = Node & {
  __typename?: 'Seller';
  createdAt: Scalars['DateTime']['output'];
  customFields?: Maybe<Scalars['JSON']['output']>;
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type SendTestEmailInput = {
  template: Scalars['String']['input'];
  to: Scalars['String']['input'];
};
// ...
export type TemplatePhraseInput = {
  id?: InputMaybe<Scalars['ID']['input']>;
  key: Scalars['String']['input'];
  templateCode: Scalars['String']['input'];
  value: Scalars['String']['input'];
};
// ...
export type UpdateEmailTemplateInput = {
  attachFile?: InputMaybe<Scalars['Boolean']['input']>;
  description?: InputMaybe<Scalars['String']['input']>;
  enabled?: InputMaybe<Scalars['Boolean']['input']>;
  id: Scalars['ID']['input'];
  name?: InputMaybe<Scalars['String']['input']>;
  phrases?: InputMaybe<Array<TemplatePhraseInput>>;
  template?: InputMaybe<Scalars['String']['input']>;
};